requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.1.3
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.14.0
Markdown==3.4.1
packaging==21.3
pyparsing==3.0.9
pytz==2022.6
redis==4.3.4
sqlparse==0.4.3
wrapt==1.14.1
